+++ CHANGELOG.md
@@ -15,9 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Refactored ARCO datasource to use asyncio
+
 ### Deprecated
 
 ### Removed
+
+- Removed tp06 from ARCO, use WB2 instead
 
 ### Fixed